+++ src/graphics/selection_list.rs
@@ -2,8 +2,7 @@
 use iced_graphics::{backend, Backend, Color, Point, Primitive, Renderer};
 use iced_native::{mouse, HorizontalAlignment, Rectangle, VerticalAlignment};
 
-use crate::native::overlay::list;
-pub use crate::native::selection_list::{self, State};
+pub use crate::native::selection_list::{self, list, State};
 pub use crate::style::selection_list::{Style, StyleSheet};
 
 /// A widget allowing the selection of a single value from a list of options.
